[OP]Fix adamw not registered into AllKernels
